ONNX: create constant nodes for used initializers
Collect node inputs not connected either to node outputs or graph
inputs and search for initializers for them. In case one found,
create tensor constant node for it and connect the corresponding
input to it. It doesn't modify the existing graph connection process
and is used only as a fallback branch instead of exception
previously being thrown.