ONNX: add BatchNormalization operator (#136)
* ONNX: fix formatting and indentation issues

* ONNX: extract method to add tensor constant node

It used to be in the constant operator processing code, but turned
out to be very useful in processing of other operators as well.

* ONNX: remove `std::string_view` in node collections

To make the value passing more streamlined and avoid any potential
problems of original string deallocation somewhere inside Protobuf
structures.

* ONNX: create constant nodes for used initializers

Collect node inputs not connected either to node outputs or graph
inputs and search for initializers for them. In case one found,
create tensor constant node for it and connect the corresponding
input to it. It doesn't modify the existing graph connection process
and is used only as a fallback branch instead of exception
previously being thrown.

* ONNX: add `BatchNormalization` operator