ONNX: fix model import when `graph.value_info` is empty
Fix `Conv` and `ConvTranspose` ops conversion, which have become too
curious on output information and lead to broken ONNX import for the
models which don't contain `ValueInfo` for inputs and outputs of all
the nodes in the `value_info` field of the graph.

For `Conv` op input data type is used for output and weights, which is
a must according to
[ONNX spec](https://github.com/onnx/onnx/blob/main/docs/Operators.md#type-constraints-26).
Output shape is not used and is removed.

For `ConvTranspose` `generate_output_shape()` function is restored to
initialize output shape in case of absence of the corresponding
attribute (otherwise it's read from the attribute value).

Concerns #310, #549, #103.